All vehicles: increase manual throttle resolution (original branch manual_res)
commit 99a3a42f9da3e63bee4b8163804819493cbcb2b5
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:27:33 2021 +0200

    AP_OSD_Screen: draw_throttle: round float throttle

commit 6f61077e63f5672d8c956433631e2516f4f2575c
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:14:33 2021 +0200

    GCS_MAVLink: change return of vfr_hud_throttle() from int16 to float

commit 99ec0bb80bcc248ce8271e3e0f6cc09a10178149
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:13:58 2021 +0200

    GCS_MAVLINK_Rover: change return of vfr_hud_throttle() from int16 to float

commit 8e740bc0906556983bc83889bca20b74e4644eb4
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:13:39 2021 +0200

    GCS_MAVLINK_Blimp: change return of vfr_hud_throttle() from int16 to float

commit fe3a05f9c882e1e9118b63da813f6223a81665e2
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:13:16 2021 +0200

    GCS_MAVLINK_Sub: change return of vfr_hud_throttle() from int16 to float

commit b8b494b51bc9793bda354bf7ea1345141736f08b
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:12:26 2021 +0200

    Plane: change return of throttle_percentage() from int8 to float

commit 12f0069160e2472a725c0fb749391849548b7c36
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:11:17 2021 +0200

    GCS_MAVLINK_Plane: change return of vfr_hud_throttle() from int16 to float

commit b75b4badcc22951c5032bda229de9159fdb25e22
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Thu Oct 21 03:10:36 2021 +0200

    GCS_MAVLINK_Copter: change return of vfr_hud_throttle() from int16 to float

commit d2a518c1593f52051e5f3639120c01432e560375
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:29:17 2021 +0200

    AP_Arming: changed needed after making scaled throttle a float

commit e2c58978a850a7844bb4170b3e53879c8a25b619
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:28:33 2021 +0200

    Rover: change throttle and steering input logging to float

commit faac500b20524023638c0bae528e6973249ed577
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:28:04 2021 +0200

    Blimp: necessary changed after making scaled throttle a float

commit 371dcad527ec22059c306856dcd46df50d8922e8
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:27:44 2021 +0200

    Plane: necessary changed after making scaled throttle a float

commit c8cc9e8f7bc2528051b20355b934380371fc006a
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:27:15 2021 +0200

    Copter: changes necessary after making scaled throttle channel a float

commit 2c2adf4a8030e816ddae1a4aeeadf31b4fea0e75
Author: Michel Pastor <shellixyz@gmail.com>
Date:   Tue Oct 19 22:23:22 2021 +0200

    RC_Channel: return scaled channel value as float

+++ ArduPlane/Plane.h
@@ -1053,7 +1053,7 @@ private:
     void check_short_failsafe();
     void startup_INS_ground(void);
     bool should_log(uint32_t mask);
-    int8_t throttle_percentage(void);
+    float throttle_percentage(void);
     void notify_mode(const Mode& mode);
 
     // takeoff.cpp
